examples/example-ai-gemini/pyproject.toml
[project]
name = "example-ai-gemini"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "opentelemetry-instrumentation-google-generativeai>=0.24.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp-proto-http>=1.30.0",
]
